requirements.txt
asgiref==3.5.0
attrs==21.4.0
Django==4.0.3
djangorestframework==3.13.1
gunicorn==20.1.0
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyparsing==3.0.8
pytest==7.1.1
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
